TIP-1241: Compute missing required attributes for PEF
