+++ tests/back/Pim/Enrichment/Specification/Component/Product/Completeness/CompletenessCalculatorSpec.php
@@ -104,59 +104,4 @@ class CompletenessCalculatorSpec extends ObjectBehavior
             new ProductCompletenessWithMissingAttributeCodesCollection(5, [])
         );
     }
-
-    function it_calculates_the_completeness_of_an_entity_with_family(
-        GetCompletenessProductMasks $getCompletenessProductMasks,
-        GetRequiredAttributesMasks $getRequiredAttributesMasks,
-        ProductModelInterface $entityWithFamily
-    ) {
-        $family = new Family();
-        $family->setCode('accessories');
-        $entityWithFamily->getFamily()->willReturn($family);
-        $entityWithFamily->getCode()->willReturn('my_bag');
-        $entityWithFamily->getId()->willReturn(42);
-        $values = new WriteValueCollection();
-        $entityWithFamily->getValues()->willReturn($values);
-
-        $requiredAttributesMasksPerChannelAndLocale = [
-            new RequiredAttributesMaskForChannelAndLocale(
-                'ecommerce',
-                'en_US',
-                ['name-ecommerce-en_US', 'view-ecommerce-en_US']
-            ),
-            new RequiredAttributesMaskForChannelAndLocale(
-                '<all_channels>',
-                '<all_locales>',
-                ['desc-<all_channels>-<all_locales>']
-            ),
-        ];
-        $requiredAttributesMask = new RequiredAttributesMask(
-            'accessories', $requiredAttributesMasksPerChannelAndLocale
-        );
-        $getRequiredAttributesMasks->fromFamilyCodes(['accessories'])->willReturn(
-            ['accessories' => $requiredAttributesMask]
-        );
-
-        $productCompletenessMask = new CompletenessProductMask(
-            42, 'my_bag', 'accessories', [
-                'name-ecommerce-en_US',
-                'name-ecommerce-fr_FR',
-                'desc-<all_channels>-<all_locales>',
-                'price-tablet-fr_FR',
-                'size-ecommerce-en_US',
-            ]
-        );
-        $getCompletenessProductMasks->fromValueCollection(42, 'my_bag', 'accessories', $values)->willReturn(
-            $productCompletenessMask
-        );
-
-        $this->fromEntityWithFamily($entityWithFamily)->shouldBeLike(
-            new ProductCompletenessWithMissingAttributeCodesCollection(
-                42, [
-                    new ProductCompletenessWithMissingAttributeCodes('ecommerce', 'en_US', 2, [1 => 'view']),
-                    new ProductCompletenessWithMissingAttributeCodes('<all_channels>', '<all_locales>', 1, []),
-                ]
-            )
-        );
-    }
 }
